+++ NEWS.md
@@ -14,6 +14,7 @@ sudo make install
 - Add number of employees and use case to user profile ([#14966](https://github.com/CartoDB/cartodb/pull/14966))
 
 ### Bug fixes / enhancements
+- Stricter email domain validation ([#15030](https://github.com/CartoDB/cartodb/pull/15030))
 - Add more columns to oauth_app ([#15015](https://github.com/CartoDB/cartodb/issues/15015))
 - Redirect viewer users to shared visualizations page, and show shared visualizations in Home ([CartoDB/support#2032](https://github.com/CartoDB/support/issues/2032))
 - Fix user presenter ([#15033](https://github.com/CartoDB/cartodb/pull/15033))